pyproject.toml
[project]
name = "apteka-notify"
version = "0.1.0"
description = "Автономная система уведомлений для аптеки"
requires-python = ">=3.11"
readme = "README.md"

dependencies = [
    "aiogram>=3.4",
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "html2text>=2024.2",
    "pdfplumber>=0.10",
    "python-docx>=1.1",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "aiosqlite>=0.20",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
